+++ scilab2py/tests/test_usage.py
@@ -9,7 +9,7 @@ from scilab2py import Scilab2Py, Scilab2PyError, scilab
 from scilab2py.utils import Struct
 
 THIS_DIR = os.path.abspath(os.path.dirname(__file__))
-scilab.close()
+scilab.exit()
 
 
 class BasicUsageTest(test.TestCase):
@@ -20,51 +20,49 @@ class BasicUsageTest(test.TestCase):
         self.sci.getd(THIS_DIR)
 
     def tearDown(self):
-        self.sci.close()
+        self.sci.exit()
 
     def test_run(self):
         """Test the run command
         """
-        self.sci.run('y=ones(3,3)')
-        y = self.sci.get('y')
+        self.sci.eval('y=ones(3,3)')
+        y = self.sci.pull('y')
         desired = np.ones((3, 3))
         test.assert_allclose(y, desired)
-        self.sci.run('x = mean([[1, 2], [3, 4]])')
-        x = self.sci.get('x')
+        self.sci.eval('x = mean([[1, 2], [3, 4]])')
+        x = self.sci.pull('x')
         self.assertEqual(x, 2.5)
-        self.assertRaises(Scilab2PyError, self.sci.run, '_spam')
+        self.assertRaises(Scilab2PyError, self.sci.eval, '_spam')
 
-    def test_call(self):
-        """Test the call command
+    def test_dynamic_functions(self):
+        """Test some dynamic functions
         """
-        out = self.sci.call('ones', 1, 2)
+        out = self.sci.ones(1, 2)
         assert np.allclose(out, np.ones((1, 2)))
-        U, S, V = self.sci.call('svd', [[1, 2], [1, 3]])
+        U, S, V = self.sci.svd([[1, 2], [1, 3]])
         assert np.allclose(U, ([[-0.57604844, -0.81741556],
                            [-0.81741556, 0.57604844]]))
         assert np.allclose(S,  ([[3.86432845, 0.],
                            [0., 0.25877718]]))
         assert np.allclose(V,  ([[-0.36059668, -0.93272184],
                            [-0.93272184, 0.36059668]]))
-        out = self.sci.call('roundtrip.sci', 1)
+        out = self.sci.roundtrip(1)
         self.assertEqual(out, 1)
-        fname = os.path.join(THIS_DIR, 'roundtrip.sci')
-        out = self.sci.call(fname, 1)
         self.assertEqual(out, 1)
-        self.assertRaises(Scilab2PyError, self.sci.call, '_spam')
+        self.assertRaises(Scilab2PyError, self.sci.eval, '_spam')
 
-    def test_put_get(self):
+    def test_push_pull(self):
         """Test putting and getting values
         """
-        self.sci.put('spam', [1, 2])
-        out = self.sci.get('spam')
+        self.sci.push('spam', [1, 2])
+        out = self.sci.pull('spam')
         assert np.allclose(out, np.array([1, 2]))
-        self.sci.put(['spam', 'eggs'], ['foo', [1, 2, 3, 4]])
-        spam, eggs = self.sci.get(['spam', 'eggs'])
+        self.sci.push(['spam', 'eggs'], ['foo', [1, 2, 3, 4]])
+        spam, eggs = self.sci.pull(['spam', 'eggs'])
         self.assertEqual(spam, 'foo')
         assert np.allclose(eggs, np.array([[1, 2, 3, 4]]))
-        self.assertRaises(Scilab2PyError, self.sci.put, '_spam', 1)
-        self.assertRaises(Scilab2PyError, self.sci.get, '_spam')
+        self.assertRaises(Scilab2PyError, self.sci.push, '_spam', 1)
+        self.assertRaises(Scilab2PyError, self.sci.pull, '_spam')
 
     def test_dynamic(self):
         """Test the creation of a dynamic function
@@ -83,13 +81,14 @@ class BasicUsageTest(test.TestCase):
         """Test opening and closing the Scilab session
         """
         sci_ = Scilab2Py()
-        sci_.close()
-        self.assertRaises(Scilab2PyError, sci_.put, names=['a'],
+        sci_.exit()
+        self.assertRaises(Scilab2PyError, sci_.push, name=['a'],
                           var=[1.0])
         sci_.restart()
-        sci_.put('a', 5)
-        a = sci_.get('a')
+        sci_.push('a', 5)
+        a = sci_.pull('a')
         assert a == 5
+        sci_.exit()
 
     def test_struct(self):
         """Test Struct construct
@@ -111,17 +110,17 @@ class BasicUsageTest(test.TestCase):
         """Make sure a syntax error in Scilab throws an Scilab2PyError
         """
         sci = Scilab2Py()
-        self.assertRaises(Scilab2PyError, sci._eval, "a='1")
+        self.assertRaises(Scilab2PyError, sci.eval, "a='1")
         sci = Scilab2Py()
-        self.assertRaises(Scilab2PyError, sci._eval, "a=1+*3")
+        self.assertRaises(Scilab2PyError, sci.eval, "a=1+*3")
 
-        sci.put('a', 1)
-        a = sci.get('a')
+        sci.push('a', 1)
+        a = sci.pull('a')
         self.assertEqual(a, 1)
 
     def test_scilab_error(self):
         sci = Scilab2Py()
-        self.assertRaises(Scilab2PyError, sci.run, 'a = ones2(1)')
+        self.assertRaises(Scilab2PyError, sci.eval, 'a = ones2(1)')
 
     def test_context_manager(self):
         '''Make sure Scilab2Py works within a context manager'''