+++ HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 
 0.2.0 (2014-08-14)
 ++++++++++++++++++
+- Streamline API to mirror Oct2Py 2.0.0
 - Python 3 support
 - Bug fixes and usability improvements.